rust: capnp: import crate
This is a subset of the Rust `capnp` crate,
version capnpc-v0.18.0, licensed under "MIT", from:

    https://github.com/capnproto/capnproto-rust/tree/capnpc-v0.18.0/capnp/src

The files are copied as-is, with no modifications whatsoever
(not even adding the SPDX identifiers).

For copyright details, please see:

    https://github.com/capnproto/capnproto-rust/blob/capnpc-v0.18.0/LICENSE

The next patches modify these files as needed for use within the kernel.
This patch split allows reviewers to double-check the import and to
clearly see the differences introduced.

The following script may be used to verify the contents:

    for path in $(cd rust/capnp/ && find . -type f -name '*.rs'); do
        curl --silent --show-error --location \
            https://github.com/capnproto/capnproto-rust/raw/capnpc-v0.18.0/capnp/src/$path \
            | diff --unified rust/capnp/$path - && echo $path: OK
    done

Signed-off-by: Ariel Miculas <amiculas@cisco.com>